Distribute AwesomeBar composable implementation over multiple files.
